Added new mode to generate heightmap data per-polygon.
This is much faster than the old way (interpolation).
The catch is that it gives very noticable 'steps' in the heightmap at
polygon boundaries.

A side effect of this commit is adding 2 new functions, allowing users to
take a position in 2D space and return the polygon encompassing that point.