Moved Biome processing out of the task graph and fixed a bug with heightmap ranges.
The change to biome processing should close issue #12, since GameplayTags don't play nice on multiple threads. This takes the GameplayTag system off of the multithreaded approach and instead runs biome determination as a post-processing step on the game thread. This will have a slight decrease in performance, but hopefully it should no longer crash randomly on some systems.

Additionally, the Elevation property on each heightmap pixel (which was assumed to be between 0 and 1) was actually being set to values between 0 and ~25000. This was creating an integer overflow when writing a heightmap to disk, as the absurdly high Elevation values were being multiplied by 255 and cast to an 8-bit int. The 100+ value difference between neighboring points were creating a psychadelic effect (as seen in issue #16) and should now be fixed.