Refactored all basic array data structures.
These new arrays can be accessed with custom structures like FPointIndex or FTriangleIndex, which can be treated like a standard SIZE_T value.
The reason for this change is to help differentiate when a value is a Side vs. a Triangle or Point, as before they were all int32 values.
SIZE_T was chosen as it's Unreal's way of representing the C++ size_t value, which is used by the Delaunator library when generating the map.